Rename FactoryRegistry to WidgetRegistry
