+++ src/WidgetRegistry.ts
