+++ src/main.ts
@@ -1,7 +1,7 @@
 export * from './animations/cssTransitions';
 export * from './customElements';
 export * from './d';
-export * from './FactoryRegistry';
+export * from './WidgetRegistry';
 export * from './interfaces';
 export * from './mixins/FormLabel'
 export * from './mixins/I18n';
